fix: rework bottom alert box
