+++ apps/ledger-live-mobile/src/families/tron/ConfirmationPostAlert.tsx
@@ -1,0 +1,24 @@
+import type { Account } from "@ledgerhq/types-live";
+import React from "react";
+import { Trans } from "react-i18next";
+import Alert from "../../components/Alert";
+import { urls } from "../../config/urls";
+
+type Props = {
+  mainAccount: Account;
+};
+
+export default function ReceiveConfirmationAlertPost({ mainAccount }: Props) {
+  return (
+    <>
+      {mainAccount.operationsCount === 0 ? (
+        <Alert
+          type="warning"
+          learnMoreUrl={urls.errors.TronSendTrc20ToNewAccountForbidden}
+        >
+          <Trans i18nKey="tron.receive.newAddressTRC20" />
+        </Alert>
+      ) : null}
+    </>
+  );
+}
