+++ apps/ledger-live-mobile/src/screens/ReceiveFunds/03-Confirmation.tsx
@@ -30,6 +30,8 @@ import { ScreenName } from "../../const";
 import { track, TrackScreen } from "../../analytics";
 import PreventNativeBack from "../../components/PreventNativeBack";
 import byFamily from "../../generated/Confirmation";
+import byFamilyPostAlert from "../../generated/ConfirmationPostAlert";
+
 import { ReceiveFundsStackParamList } from "../../components/RootNavigator/types/ReceiveFundsNavigator";
 import {
   BaseComposite,
@@ -204,6 +206,17 @@ function ReceiveConfirmationInner({
     }
   }
 
+  let CustomConfirmationAlert;
+  if (
+    currency.type === "CryptoCurrency" &&
+    Object.keys(byFamilyPostAlert).includes(currency.family)
+  ) {
+    CustomConfirmationAlert =
+      currency.type === "CryptoCurrency"
+        ? byFamilyPostAlert[currency.family as keyof typeof byFamilyPostAlert]
+        : null;
+  }
+
   return (
     <Flex flex={1} mb={9}>
       <PreventNativeBack />
@@ -329,6 +342,9 @@ function ReceiveConfirmationInner({
             })}
           </Text>
         </Flex>
+        {CustomConfirmationAlert && (
+          <CustomConfirmationAlert mainAccount={mainAccount} />
+        )}
       </NavigationScrollView>
       <Flex m={6}>
         {isToastDisplayed ? (
